Issues resolve attempt#1

+++ app/src/main/java/com/example/kreaz/data/cart.kt
@@ -3,7 +3,6 @@ package com.example.kreaz.data
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.text.NumberFormat
 
 @Entity
 data class cart(
@@ -12,7 +11,7 @@ data class cart(
     @ColumnInfo(name = "name")
     val itemName: String = "add items",
     @ColumnInfo(name = "price")
-    val itemPrice: Float = 0.0.toFloat(),
+    val itemPrice: Float = 0f,
     @ColumnInfo(name = "quantity")
     val quantity: Int = 0,
 )
